changed git fetch from nobranch/AUTOREV to main branch (#79)
* changed from nobranch/AUTOREV to main branch

This fixes a recent bug introduced over the weekend which results in a bitbake expansion error. 
Specifically the error is the following,
Fetcher failure: Unable to resolve 'master' in upstream git repository in git ls-remote output for github.com/Azure/hyperlocal-windows

* nobranch/AUTOREV to main branch in iotedge-daemon

This fixes a recent bug introduced over the weekend which results in a bitbake expansion error. 
Specifically the error is the following,
Fetcher failure: Unable to resolve 'master' in upstream git repository in git ls-remote output for github.com/Azure/hyperlocal-windows