Fix rust library location
The path to rust lib dir was hardcoded into the aziotd include file.
Because the path was not matching the one used on rust side it, the
services using shared rust libraries were failing and could not
find the shared library.

This commit replaces the hardcoded value by the variable which tells rust
(in this case cargo) where to place the compiled libs.
This way, in case of future evolution of the rustc lib path, the up to
date one will be used.

Signed-off-by: Clement Ramirez <clement.ramirez@actia.fr>